+++ vue-study/src/hooks/useSum.ts
@@ -1,0 +1,11 @@
+import {ref} from "vue";
+
+export default function (){
+    let sum = ref(0);
+
+    function add(){
+        sum.value += 1;
+    }
+
+    return {sum,add}
+}
